+++ src/components/Header.js
@@ -1,17 +1,24 @@
 import { Logo } from "./SignIn";
 import "../css/header.css";
 
-export default function Header() {
+export default function Header({ showing, setShowing }) {
+    const buttonsText = ["home", "games", "scoreboards", "challenge"];
+
     return (
         <div className="header">
             <Logo />
             <div className="header-text">
                 <h1>Kings Corner</h1>
                 <div className="navigation">
-                    <button>dashboard</button>
-                    <button>games</button>
-                    <button>scoreboards</button>
-                    <button>challenge</button>
+                    {buttonsText.map((text) => (
+                        <button
+                            key={`nav button ${text}`}
+                            onClick={() => setShowing(text)}
+                            disabled={showing === text}
+                        >
+                            {text}
+                        </button>
+                    ))}
                 </div>
             </div>
         </div>
