+++ src/css/dashboard.scss
@@ -1,6 +1,11 @@
 .dashboard {
     text-align: center;
 
+    h2.greeting {
+        font-size: 28px;
+        font-weight: 400;
+    }
+
     .pick-opponents {
         width: 200px;
         margin: auto;
