+++ src/css/header.scss
@@ -3,7 +3,7 @@
     display: flex;
     align-items: center;
     gap: 20px;
-    padding: 10px;
+    padding: 10px 10px 10px 15px;
 
     .logo {
         height: 50px;
@@ -31,6 +31,11 @@
                 font-size: 16px;
                 padding: 0;
                 text-decoration: underline;
+
+                &:disabled {
+                    color: #aaa;
+                    text-decoration: none;
+                }
             }
         }
     }
